Refactored TH.hs into multiple files.
